Changes to ensure decentTree can build directly under Visual Studio (and CMake) without needing you to install (or use) the CLang compiler, on Windows. 1.   Renamed utils/io.h to utils/safe_io.h (there were name       collisions that made look-up of _isatty problematic       under the Visual Studio compiler).       This forced #include directive updates in       (a) alignment/alignment.cpp       (b) alignment/superalignment.cpp       (c) main/timetree.cpp       (d) utils/checkpoint.cpp       (e) utils/distancematrix.cpp       It also forced an update to utils/CMakeLists.txt. 2.   Added more versions of progress_display member       functions (so, fewer conversion warnings or explicit       conversions needed, in MSVC builds).  Ironically this       forced some additional casts (from int to double) in:       (a) alignment/alignment.cpp       (b) alignment/alignmentsummary.cpp (literals)       (c) utils/decenttree.cpp 3. Redeclared a lot of size_t variables and members, used     in classes involved in distance matrix algorithms, to be     intptr_t.  The issue was that Visual Studio's version of     OMP doesn't like you to used *unsigned* types in     "#omp parallel for" loops. 4. Declared cell_type types for Matrix and FlatMatrix     (so that loadDistanceMatrixInto() can avoid conversion      warnings!). 5. FlatMatrix::getMaxSeqNameLength() (which already      returned size_t) does its calculations in size_t rather      than int. 6.  gzstreambuf::open calls fopen_s rather than fopen,      under Visual Studio. 7.  gzstreambuf::underflow() uses intptr_t in place of long      in a "pointer-difference" calculation.  Its use of long was      potentially a scaling fault. 8.  casts to and from T (the type of the distance matrix)      are now explicit, in nj.h. 9.  BIONJMatrix::chooseLambda uses intptr_t rather than      size_t.  As does  BoundingMatrix::purgeRow(),      UPGA_Matrix::cluster (and all the other overrides      of cluster) (including even the one in       tree/parsimonyjoining.cpp). 10. isStandardOutputATerminal (in operatingsystem.cpp)       references _fileno rather than fileno in MSVC builds       (it avoids a warning about fileno being deprecated). 11. BoundingMatrix::clusterToRow is now a vector of       intptr_t rather than int (avoids conversion warnings). 12. utils/timeutil.h includes a different config file in       DECENT_TREE builds. 13. utils/timeutil.h's gettimeofday uses _ftime64_s       in MSVC builds (this might need changing; it might       have to use _ftime32_s in 32 bit Windows builds). 14. Got rid of a hardcoded float declaration in upgma.h       (in UPGMA_Matrix::getRowMinima()), and some       hardcoded double declarations too (those were in       UPGMA_Matrix::cluster()). 15. In utils/bionj.cpp disabled CRT warnings, and float       conversion warning errors.  Set Qmin to something       that didn't cause a floating-point overflow (!), in       Best_pair().

+++ alignment/alignment.cpp
@@ -20,7 +20,7 @@
 #include "utils/timeutil.h"        //for getRealTime()
 #include "utils/progress.h"        //for progress_display
 #include "utils/hammingdistance.h" //for sumForUnknownCharacters
-#include "utils/io.h"              //for safeGetLine
+#include "utils/safe_io.h"         //for safeGetLine
 #include "alignmentsummary.h"
 
 #include <Eigen/LU>
@@ -1868,8 +1868,8 @@ bool Alignment::constructPatterns(int nseq, int nsite,
     patternInfo.resize((nsite + (step-1)) / step);
     progress_display* progress_here = nullptr;
     if (progress==nullptr && !isShowingProgressDisabled) {
-        progress_here = new progress_display( nsite, "Constructing alignment",
-                                                               "examined", "site");
+        progress_here = new progress_display( (double) nsite, "Constructing alignment",
+                                              "examined", "site");
         progress = progress_here;
     }
 #ifdef _OPENMP
@@ -1932,7 +1932,7 @@ bool Alignment::constructPatterns(int nseq, int nsite,
             info.isAllGaps = pat.isAllGaps(STATE_UNKNOWN);
         }
         if (progress!=nullptr) {
-            progress->incrementBy(step);
+            progress->incrementBy((double)step);
         }
     }
     if (progress_here!=nullptr) {
@@ -1946,7 +1946,7 @@ bool Alignment::constructPatterns(int nseq, int nsite,
     //2. Now handle warnings and errors, and compress patterns, sequentially
 
     if (progress==nullptr && !isShowingProgressDisabled) {
-        progress_here = new progress_display(nsite, "Compressing patterns",
+        progress_here = new progress_display((double)nsite, "Compressing patterns",
                                              "processed", "site");
         progress = progress_here;
     }
@@ -1985,7 +1985,7 @@ bool Alignment::constructPatterns(int nseq, int nsite,
             }
         }
         if (progress!=nullptr) {
-            progress->incrementBy(step);
+            progress->incrementBy((double)step);
         }
     }
     resize(w);
@@ -3231,7 +3231,7 @@ void Alignment::extractSubAlignment(Alignment *aln, IntVector &seq_id, int min_t
         }
         if (progress!=nullptr) {
             if (siteMod == 100 ) {
-                progress->incrementBy(100);
+                progress->incrementBy(100.0);
                 siteMod  = 0;
             }
             ++siteMod;
